+++ README.md
@@ -1,10 +1,6 @@
 
 # mdx-go
 
-- No style/layout components
-- Lazy loaded LiveCode
-- [ ] Directory view
-
 ![](https://s3.amazonaws.com/jxnblk/mdx-go-24.gif)
 
 :zap: Lightning fast [MDX][]-based dev server for progressive documentation
@@ -25,8 +21,8 @@ npm i -g mdx-go
 - :atom_symbol: Import and use React components
 - :file_folder: File-system based routing
 - :triangular_ruler: Customizable layouts
-- :woman_singer: Support for [styled-components][] & [emotion][]
 - :globe_with_meridians: Export as static HTML
+- :woman_singer: Support for [styled-components][] & [emotion][]
 - :unlock: Avoid lock-in and easily migrate to other MDX-based tools
 
 [badge]: https://flat.badgen.net/travis/jxnblk/mdx-go
