+++ examples/dev/index.mdx
@@ -8,11 +8,6 @@ import {
 
 export const files = require.context('..', true, /\.example\.mdx?$/, 'lazy')
 
-export const xRoot = props =>
-  <div>
-    {props.children}
-  </div>
-
 # Dev Example
 
 This example can be run from the root directory as a development environment for the core library.
