+++ lib/client/Keyboard.js
@@ -19,8 +19,6 @@ export class Keyboard extends React.Component {
       case 'j':
         this.goto(1)
         break
-      default:
-        console.log(e.key, e.altKey)
     }
   }
