+++ lib/client/components.js
@@ -3,3 +3,4 @@ export { Link } from './Link'
 export { ComponentProvider } from './ComponentProvider'
 export { withComponents } from './withComponents'
 export { StyleProvider } from './StyleProvider'
+export { lazyComponent } from './lazyComponent'
