+++ lib/client/lazyComponent.js
@@ -24,7 +24,6 @@ export const lazyComponent = (loader, loading) => {
 
       load.then(({
         default: Component,
-        // ...exports
       }) => {
         this.setState({ Component })
       })
