+++ lib/client/scope.js
@@ -2,12 +2,29 @@ import React from 'react'
 import { Link } from './Link'
 import lazyComponent from './lazyComponent'
 
-const pre = lazyComponent(() => import('./LiveCode'))
+const pre = lazyComponent(() => import('mdx-live'))
+
+const heading = Tag => ({ id, children, ...props }) =>
+  <Tag id={id} {...props}>
+    <a
+      href={'#' + id}
+      style={{
+        textDecoration: 'none',
+        color: 'inherit'
+      }}
+      children={children}
+    />
+  </Tag>
 
 export const scope = {
   a: Link,
-  pre: props => props.children,
-  code: pre,
+  pre,
+  h1: heading('h1'),
+  h2: heading('h2'),
+  h3: heading('h3'),
+  h4: heading('h4'),
+  h5: heading('h5'),
+  h6: heading('h6'),
 }
 
 export default scope
