+++ package.json
@@ -50,14 +50,14 @@
     "is-absolute-url": "^2.1.0",
     "lodash.get": "^4.4.2",
     "lodash.sortby": "^4.7.0",
-    "mdx-style": "^1.0.0-3",
+    "mdx-live": "^2.0.0-alpha.2",
+    "mdx-style": "^2.0.0-alpha.2",
     "meow": "^5.0.0",
     "pkg-conf": "^2.1.0",
     "prop-types": "^15.6.2",
     "react": "^16.5.2",
     "react-dev-utils": "^5.0.2",
     "react-dom": "^16.5.2",
-    "react-live": "^1.11.0",
     "read-pkg-up": "^4.0.0",
     "remark-autolink-headings": "^5.0.0",
     "remark-emoji": "^2.0.1",
